perf: move `lattice` from `anna_api` to `anna`
