+++ api/Cargo.toml
@@ -8,6 +8,6 @@ license = "Apache-2.0"
 
 [dependencies]
 serde = { version = "1.0.126", features = ["derive", "rc"] }
-chrono = { version = "0.4.19", features = ["serde"] }
+chrono = { version = "=0.4.19", features = ["serde"] }
 eyre = "0.6.5"
 serde_json = "1.0.64"
