+++ src/lattice/bool.rs
