+++ src/lattice/causal/mod.rs
