+++ src/lattice/causal/single_key_causal.rs
