+++ src/lattice/max.rs
